+++ BoEG/Assets/Scripts/Framework/Core/Modules/Ability/AbilityView.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 21689630f7494403abd7f7b666f4bec2
+timeCreated: 1612909352
