+++ BoEG/Assets/Scripts/Framework/Core/Modules/Ability/IAbilityView.cs
@@ -1,4 +1,3 @@
-using MobaGame.Entity.Abilities;
 using UnityEngine;
 
 namespace MobaGame.Framework.Core.Modules.Ability
@@ -13,31 +12,5 @@ namespace MobaGame.Framework.Core.Modules.Ability
 
         IToggleableAbilityView Toggleable { get; }
 
-    }
-
-    public class AbilityView : IAbilityView
-    {
-        public Sprite Icon { get; set; }
-
-
-        public ICooldownAbilityView Cooldown { get; set; }
-        public IStatCostAbilityView StatCost { get; set; }
-
-        public IToggleableAbilityView Toggleable { get; set; }
-    }
-
-    public class SimpleAbilityView : IAbilityView
-    {
-        
-        public Sprite Icon { get; set; }
-
-
-        public ICooldownAbilityView Cooldown { get; set; }
-        public StatCostAbilityView StatCost { get; set; }
-        IStatCostAbilityView IAbilityView.StatCost => StatCost;
-
-        public ToggleableAbilityView Toggleable { get; set; }
-        IToggleableAbilityView IAbilityView.Toggleable => Toggleable;
-        
     }
 }
