+++ BoEG/Assets/Scripts/Framework/Core/Modules/Ability/SimpleAbilityView.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 72aa8377280e489faa7583eceed9fddf
+timeCreated: 1612909352
