+++ BoEG/Assets/Scripts/Framework/Core/Modules/Ability/SubInfo/IStatCostAbilityView.cs
@@ -6,17 +6,4 @@ namespace MobaGame.Framework.Core.Modules.Ability
         bool CanSpendCost();
 
     }
-
-    public class StatCostAbilityView : IStatCostAbilityView
-    {
-        public StatCostAbilityView(IMagicable magicable, float cost = default)
-        {
-            _magicable = magicable;
-            Cost = cost;
-        }
-
-        private readonly IMagicable _magicable;
-        public float Cost { get; set; }
-        public bool CanSpendCost() => _magicable.HasMagic(Cost);
-    }
 }
