+++ BoEG/Assets/Scripts/Framework/Core/Modules/Ability/SubInfo/StatCostAbilityView.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 65d3c775fcb1407898039c17726f9972
+timeCreated: 1612909352
